Node 6, moved lib to src
