+++ .babelrc
@@ -1,9 +1,11 @@
 {
   "presets": [
-    "es2015-node",
-    "async-to-bluebird"
+    "es2015-node"
   ],
   "plugins": [
-    "transform-object-rest-spread"
+    "transform-es2015-parameters",
+    "transform-es2015-destructuring",
+    "transform-object-rest-spread",
+    "transform-async-to-generator"
   ]
 }
