+++ .nvmrc
@@ -1,1 +1,1 @@
-5.10.0
+6
