+++ .travis.yml
@@ -1,4 +1,4 @@
 language: node_js
 node_js:
-  - "5.10.0"
+  - "6"
 script: npm run travis-test
