+++ package.json
@@ -4,7 +4,7 @@
   "description": "Turn SoundCloud profiles into podcasts.",
   "main": "index.js",
   "engines": {
-    "node": "5.10.0"
+    "node": "6.x"
   },
   "author": "L33T KR3W",
   "contributors": [
@@ -16,27 +16,29 @@
   "homepage": "https://github.com/L33T-KR3W/soundcasts-server",
   "bugs": "https://github.com/L33T-KR3W/soundcasts-server/issues",
   "scripts": {
-    "prestart": "rm -rf build && mkdir build && babel lib --out-dir build --source-maps --optional runtime",
+    "transpile": "BABEL_ENV=development babel src --out-dir build --source-maps",
+    "prestart": "rm -rf build && mkdir build && npm run transpile",
     "start": "NODE_ENV=production node index.js",
-    "cover": "NODE_ENV=test babel-node node_modules/.bin/babel-istanbul cover test/index.js",
-    "test": "NODE_ENV=test babel-node test/index.js | tap-spec",
+    "cover": "BABEL_ENV=development NODE_ENV=test babel-node node_modules/.bin/babel-istanbul cover test/index.js",
+    "test": "BABEL_ENV=development NODE_ENV=test babel-node test/index.js | tap-spec",
     "travis-test": "npm run cover && ((cat coverage/lcov.info | coveralls) || exit 0)"
   },
   "dependencies": {
-    "babel-cli": "6.6.5",
-    "babel-plugin-transform-object-rest-spread": "6.6.5",
-    "babel-preset-async-to-bluebird": "1.0.0",
-    "babel-preset-es2015-node": "5.0.3",
-    "bluebird": "3.3.4",
+    "babel-cli": "6.x",
+    "babel-plugin-transform-async-to-generator": "6.x",
+    "babel-plugin-transform-es2015-destructuring": "6.x",
+    "babel-plugin-transform-es2015-parameters": "6.x",
+    "babel-plugin-transform-object-rest-spread": "6.x",
+    "babel-preset-es2015-node": "6.x",
     "body-parser": "1.15.0",
     "deep-extend": "0.4.1",
-    "express": "4.13.4",
+    "express": "^4.13.4",
     "js2xmlparser": "1.0.0",
     "moment": "2.12.0",
     "request": "2.69.0"
   },
   "devDependencies": {
-    "babel-istanbul": "0.7.0",
+    "babel-istanbul": "0.8.0",
     "coveralls": "2.11.9",
     "nine-track": "3.0.1",
     "tap-spec": "4.1.1",
